+++ buildSrc/src/main/kotlin/io/github/rhdunn/gradle/dsl/KotlinJvm.kt
@@ -9,13 +9,21 @@ import org.gradle.api.Project
 /**
  * Compute the publication name for the specified Java version.
  */
-fun jvmName(javaVersion: JavaVersion): String = "jvm${javaVersion.majorVersion}"
+fun jvmName(javaVersion: JavaVersion, suffix: String = ""): String =
+    "jvm${javaVersion.majorVersion}$suffix"
 
 /**
  * Access the JVM main configuration object.
  */
 val <T> NamedDomainObjectContainer<T>.jvmMain: T
     get() = findByName("jvmMain")!!
+
+/**
+ * Access the JVM main configuration object.
+ */
+fun <T> NamedDomainObjectContainer<T>.jvmMain(javaVersion: JavaVersion): T {
+    return findByName(jvmName(javaVersion, "Main"))!!
+}
 
 /**
  * Access the JVM test configuration object.