+++ buildSrc/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - A `jvmName` DSL helper method.
+- A `jvmMain` DSL helper method taking a `JavaVersion`.
 
 ### Fixed